consolidated navbar page and portcon page into one file
